Fixed Icons and Flag Bug
Flag pop-up fixed
made the map icons a bit bigger

+++ Assets/Scripts/Enemy/EnemyBaseClass.cs
@@ -25,7 +25,7 @@ public class EnemyBaseClass : MonoBehaviour
     //Base Stats
     protected int eMaxHealth;
     [SerializeField] protected int eHealth;
-    [SerializeField] protected int eSpeed;
+    [SerializeField] protected float eSpeed;
     [SerializeField] protected int eDamage;
     [SerializeField] protected string eName;
     [SerializeField] protected float eAttackSpeed;
